+++ examples/loadtest/samplecontroller/controller.go
