+++ examples/loadtest/samplestore/store.go
